Fix for https://issues.apache.org/activemq/browse/AMQ-1591
git-svn-id: https://svn.eu.apache.org/repos/asf/activemq/trunk@629639 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/store/jdbc/JDBCPersistenceAdapter.java
@@ -498,11 +498,11 @@ public class JDBCPersistenceAdapter extends DataSourceSupport implements Persist
         return 0;
     }
 
-    public int getLockKeepAlivePeriod() {
+    public long getLockKeepAlivePeriod() {
         return lockKeepAlivePeriod;
     }
 
-    public void setLockKeepAlivePeriod(int lockKeepAlivePeriod) {
+    public void setLockKeepAlivePeriod(long lockKeepAlivePeriod) {
         this.lockKeepAlivePeriod = lockKeepAlivePeriod;
     }
 }
